fsl-dpaa2-mac: resolve conflict with upstream SerDes PHY management
Downstream Linux Factory has some modifications compared to upstream.

In upstream, we support a single "phys" device tree phandle, which is
the SerDes PHY. We get it with dpaa2_mac_connect() -> of_phy_get(), put
it in mac->serdes_phy, and power-manage it through dpaa2_mac_start() and
dpaa2_mac_stop() - the same functions that also start and stop phylink.

In downstream, the code was modified in two ways:
- "phys" is now an array. The first "num-lanes" array elements are
  SerDes PHYs, and the rest are retimers.
- The SerDes PHYs are no longer managed by dpaa2-mac, but by the
  lynx-pcs. This means that dpaa2_mac_start() is no longer required to
  call phy_power_on(), because lynx_pcs_enable() does it. In turn, the
  Lynx PCS state machine is driven by phylink, so phylink_start() is
  absolutely sufficient.

So whereas in downstream, dpaa2_mac_start() is a very lightweight
wrapper over phylink_start(), and so is dpaa2_mac_stop() over
phylink_stop(), in upstream it isn't.

We currently call phylink_start() and phylink_stop() directly, but that
doesn't engage phy_power_on() and phy_power_off() in the SerDes PHY.
Because of the way in which the 28G Lynx SerDes driver treats managed vs
unmanaged lanes, it will cut power from a lane managed by the
mac-standalone driver, for which we did not call phy_power_on().
This makes the link partner see the link as down.

By simply replacing the raw phylink_start() with dpaa2_mac_start() and
phylink_stop() with dpaa2_mac_stop(), we make sure the SerDes PHY power
management is done properly also with the upstream kernel. This helps if
this driver is cherry-picked independently without the rest of the
modifications from Linux Factory, as some customers prefer to do.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>